refactor getPageSource
