+++ getPageSourceInjectionCode.js
@@ -1,0 +1,4 @@
+chrome.runtime.sendMessage({
+  sender: "mention-generator",
+  pageSource: document.documentElement.outerHTML,
+});
