+++ index.html
@@ -32,13 +32,10 @@
 
           <form>
             <h3>検索したメンバーズファイル:</h3>
-            <p class="selected-file"></p>
+            <p id="selected-file-display"></p>
             <h4>他のファイルを選択:</h4>
-            <input type="file" id="members-csv" accept=".csv" /><button
-              type="reset"
-            >
-              リセット
-            </button>
+            <input type="file" id="input-csv" accept=".csv" />
+            <button type="reset">リセット</button>
           </form>
         </details>
       </div>
